refactor: decompose alice device-capabilities.tsx
