+++ frontend/src/pages/integrations/alice/components/smart-device/components/device-skills/capabilities/on-off.tsx
@@ -1,0 +1,3 @@
+export const OnOffCapability = () => (
+  <div className="aliceDeviceSkills-colspan2"></div>
+);
